refactor:ListManangement
