--- src/components/BoardComponent.js
@@ -1,6 +1,0 @@
-class BoardComponent {
-    
-    
-}
-
-module.exports = new BoardComponent();
