--- src/components/HomeComponent.js
@@ -1,7 +1,0 @@
-class HomeComponent {
-
-
-
-}
-
-module.exports = new HomeComponent();
