+++ src/components/board/BoardListComponent.js
@@ -5,8 +5,19 @@ class BoardListComponent {
     get listAddCardButton() { return $('[data-testid="list-add-card-button"]'); }
     boardElement(boardTitle) { return $(`a[title*="${boardTitle}"]`); }
     boardItem(boardTitle) { return $(`a[aria-label*="${boardTitle}"]`); }
-    
-    
+    get inputListName() { return $('[data-testid="list-name-textarea"]'); }
+    get addListToBoardButton() { return $('[data-testid="list-composer-add-list-button"]'); }
+    listElement(listTitle) { return $(`h2[data-testid="list-name"]=${listTitle}`); }
+
+    async enterListName(listName) {
+        await this.inputListName.waitForDisplayed();
+        await this.inputListName.click();
+        await this.inputListName.setValue(listName);
+    }
+    async clickAddListToBoard() {
+        await this.addListToBoardButton.waitForDisplayed();
+        await this.addListToBoardButton.click();
+    }
 
     async validateBoardCreated(boardTitle) {
         const board = this.boardItem(boardTitle);
