+++ src/pages/BoardManagementPage.js
@@ -1,4 +1,3 @@
-const boardComponent = require('../components/BoardComponent');
 const HeaderBoardComponent = require('../components/board/headerComponent');
 const UserMenuComponent = require('../components/board/UserMenuComponent');
 const BoardListComponent = require('../components/board/BoardListComponent');
@@ -6,7 +5,6 @@ const CreateBoardComponent = require('../components/board/CreateBoardComponent')
 class BoardManagementPage {
     
     constructor() { 
-        this.board = boardComponent;
         this.headerBoard = HeaderBoardComponent;
         this.userMenu = UserMenuComponent;
         this.boardList = BoardListComponent;
