+++ src/pages/HomePage.js
@@ -1,11 +1,9 @@
-const homeComponent = require('../components/HomeComponent');
 const headerComponent = require('../components/home/HeaderComponent');
 
 
 class HomePage {
     
     constructor() {
-        this.home = homeComponent;
         this.header = headerComponent;
     }
 }
