--- src/pages/ListManagementPage.js
@@ -1,10 +1,0 @@
-const listManagementComponent = require('../components/ListManagementComponent');
-
-class ListManagementPage {
-    constructor() {
-        this.listManagement = listManagementComponent;
-    }
-}
-module.exports = new ListManagementPage();
-
-
