+++ src/step-definitions/listManagementSteps.js
@@ -1,25 +1,24 @@
 const { When, Then } = require('@wdio/cucumber-framework');
 const { expect} = require('chai');
 const BoardManagementPage = require('../pages/BoardManagementPage');
-const ListManagementPage = require('../pages/ListManagementPage');
 const testData = require('../utils/testData');
 
 let createdListTitle;
 
 When(/^the user clicks on Add a list$/, async () => {
-    await BoardManagementPage.board.clickAddList();
+    await BoardManagementPage.boardList.clickAddList();
 });
 
 When(/enters the list name "([^"]*)"$/, async (listKey) => {
     createdListTitle = testData[listKey].title;
-    await ListManagementPage.listManagement.enterListName(createdListTitle); 
+    await BoardManagementPage.boardList.enterListName(createdListTitle); 
 });
 
 When(/clicks on the Add list button$/, async () => {
-    await ListManagementPage.listManagement.clickAddListToBoard(); 
+    await BoardManagementPage.boardList.clickAddListToBoard(); 
 });
 
 Then(/the new list should appear on the board$/, async () => {
-    const list = await ListManagementPage.listManagement.listElement(createdListTitle);
+    const list = await BoardManagementPage.boardList.listElement(createdListTitle);
     expect(await list.isDisplayed()).to.be.true;
 });
